requirements.txt
requests
psycopg2-binary
gspread
oauth2client
boto3
schedule
python-dotenv
Pillow
boto3
opencv-python-headless